validate Admin

+++ app/Http/Controllers/ControllerCategories.php
@@ -7,6 +7,7 @@ use Illuminate\Foundation\Validation\ValidatesRequests;
 use Illuminate\Foundation\Bus\DispatchesJobs;
 use Validator;
 use App\categories;
+use App\smallCategories;
 use App\Http\Requests\CateRequest;
 class ControllerCategories extends Controller
 {
@@ -14,14 +15,14 @@ class ControllerCategories extends Controller
     public function getCategories(){
         $categories = categories::all();
       return view('admin.pageAdmin.categories',['categories' => $categories]);
-    } 
+    }
     public function getAddCategories(){
         return view('admin.pageAdmin.addCategories');
     }
     public function postAddCategories(CateRequest $req){
         $file_image=$req->file('cateImage')->getClientOriginalName();
         $cate=new categories;
-        $cate->name=$req->txt_CateName;
+        $cate->name=$req->txt_CateNameTest;
         $cate->image=$file_image;
         $cate->descriptions=$req->txt_CateDecription;
         $req->file('cateImage')->move('image/',$file_image);
@@ -29,19 +30,30 @@ class ControllerCategories extends Controller
         return redirect()->route('categories')->with(['flash_level'=>'success','flash_message'=>'Success !! Complete add category']);
     }
     public function getDeleteCategory($id){
-        $cate=categories::find($id);
-        $cate->delete($id);
-        return redirect()->route('categories')->with(['flash_level'=>'success','flash_message'=>'Success !! Complete delete category']);
+        $parent=smallCategories::where('id_category',$id)->count();
+        if($parent==0){
+            $cate=categories::find($id);
+            $cate->delete($id);
+            return redirect()->route('categories')->with(['flash_level'=>'success','flash_message'=>'Success !! Complete delete category']);
+        }
+        else{
+            echo"<script type='text/javascript'>
+                alert('Sorry ! You can Not Delete This Category');
+                window.location='";
+                echo route('categories');
+            echo"'</script>";
+        }
     }
     public function getEditCategory($id){
         $data=categories::find($id);
         $parent=categories::select('id','name','image','descriptions')->get()->toArray();
         return view('admin.pageAdmin.editCategory',compact('data','parent','id'));
     }
+    
     public function postEditCategory(CateRequest $req,$id){
         $cate=categories::find($id);
         $file_image=$req->file('cateImage')->getClientOriginalName();
-        $cate->name=$req->txt_CateName;
+        $cate->name=$req->txt_thienTest;
         $cate->image=$file_image;
         $cate->descriptions=$req->txt_CateDecription;
         $req->file('cateImage')->move('image/',$file_image);
